requirements.txt
flask
flask-cors
python-binance
python-dotenv